+++ anton/cli.py
@@ -296,25 +296,30 @@ def main(
     settings = AntonSettings()
     settings.resolve_workspace(folder)
 
-    # Headless mode: --prompt or --stdin
+    # Headless mode: --prompt or --stdin (mutually exclusive)
+    if prompt and stdin:
+        import sys as _sys
+        print("Error: --prompt and --stdin are mutually exclusive", file=_sys.stderr)
+        raise typer.Exit(1)
+
     headless_prompt = prompt
     if stdin:
         import sys as _sys
         headless_prompt = _sys.stdin.read().strip()
         if not headless_prompt:
-            console.print("[red]Error: no input received from stdin[/]")
+            print("Error: no input received from stdin", file=_sys.stderr)
             raise typer.Exit(1)
 
     if headless_prompt:
         # Headless mode — skip terms consent, banner, update check
         if not _has_api_key(settings):
-            console.print("[red]Error: no API key configured. Run `anton setup` first.[/]")
+            print("Error: no API key configured. Run `anton setup` first.", file=sys.stderr)
             raise typer.Exit(1)
 
         ctx.ensure_object(dict)
         ctx.obj["settings"] = settings
 
-        from anton.chat import run_headless
+        from anton.commands.headless import run_headless
 
         _ensure_workspace(settings)
         run_headless(console, settings, prompt=headless_prompt, output_format=output_format)
